requirements.txt
# YouTube API
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1

# OpenAI
openai==1.7.1

# UI
streamlit==1.29.0

# Data processing
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Additional
requests==2.31.0
